requirements.txt
flwr==1.18.0
opencv-python==4.11.0.86
numpy==2.2.6
torch==2.5.0
torchvision==0.20.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==44.0.3
filelock==3.18.0
fsspec==2025.5.1
grpcio==1.71.0
idna==3.10
iterators==0.0.2
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
protobuf==4.25.7
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
setuptools==65.5.0
shellingham==1.5.4
sympy==1.13.0
tomli==2.2.1
tomli_w==1.2.0
typer==0.12.5
typing_extensions==4.13.2
urllib3==2.4.0